Goodbye Bootstrap D:
Goodbye Bootstrap D:, it was sad but I while debuding some css I saw my bootstrap file had over 12000 rules, unsuprising, but I didn't realize the full extent on how many rules this is. My home page has in total 58 vanilla rules,  if including sweetalert2 (192 rules) it adds up to 250 rules, maybe you'll rethink using Bootstrap or something because of this, but that was the biggest reason for the switch, there were also a few tiny reasons but that's mostly it. Also fixed visual bug because of last commit adding a new NavBar, which actually was part of the plan to remove Bootstrap, so yeah.